Moves deploy to deployments module
