+++ src/prefect/__init__.py
@@ -34,7 +34,8 @@ if sys.version_info < (3, 8):
 
 
 # Import user-facing API
-from prefect.runner import Runner, serve, deploy
+from prefect.runner import Runner, serve
+from prefect.deployments import deploy
 from prefect.states import State
 from prefect.logging import get_run_logger
 from prefect.flows import flow, Flow
