+++ src/prefect/runner/__init__.py
@@ -1,1 +1,1 @@
-from .runner import Runner, serve, deploy
+from .runner import Runner, serve
