+++ src/prefect/runner/runner.py
@@ -53,7 +53,6 @@ import pendulum
 import sniffio
 from rich.console import Console, Group
 from rich.panel import Panel
-from rich.progress import Progress, SpinnerColumn, TextColumn, track
 from rich.table import Table
 
 from prefect.client.orchestration import get_client
@@ -71,7 +70,6 @@ from prefect.engine import propose_state
 from prefect.events.schemas import DeploymentTrigger
 from prefect.exceptions import (
     Abort,
-    ObjectNotFound,
 )
 from prefect.flows import Flow
 from prefect.logging.loggers import PrefectLogAdapter, flow_run_logger, get_logger
@@ -86,15 +84,8 @@ from prefect.settings import (
 )
 from prefect.states import Crashed, Pending, exception_to_failed_state
 from prefect.utilities.asyncutils import sync_compatible
-from prefect.utilities.dockerutils import (
-    PushError,
-    build_image,
-    docker_client,
-    generate_default_dockerfile,
-)
 from prefect.utilities.processutils import run_process
 from prefect.utilities.services import critical_service_loop
-from prefect.utilities.slugify import slugify
 
 if TYPE_CHECKING:
     from prefect.client.schemas.objects import FlowRun
@@ -1130,128 +1121,3 @@ async def serve(
         console.print(Panel(Group(help_message_top, table, help_message_bottom)))
 
     await runner.start()
-
-
-@sync_compatible
-async def deploy(
-    *args: RunnerDeployment,
-    work_pool_name: str,
-    image_name: str,
-    tag: Optional[str] = None,
-    dockerfile: str = "auto",
-    build_kwargs: Optional[dict] = None,
-):
-    """
-    Deploy the provided list of deployments to dynamic infrastructure via a
-    work pool.
-
-    Calling this function will build a Docker image for the deployments, push it to a
-    registry, and create each deployment via the Prefect API that will run the corresponding
-    flow on the given schedule.
-
-    Args:
-        *args: A list of deployments to deploy.
-        work_pool_name: The name of the work pool to use for these deployments.
-        image: The name of the Docker image to build, including the registry and
-            repository.
-        tag: The tag to use for the built Docker image.
-        dockerfile: The name of the Dockerfile to use for building the image. If not set
-            a Dockerfile will be generated automatically.
-        build_kwargs: Additional keyword arguments to pass to the Docker build command.
-    """
-    if not tag:
-        tag = slugify(pendulum.now("utc").isoformat())
-    full_image_name = f"{image_name}:{tag}"
-    build_kwargs = build_kwargs or {}
-
-    try:
-        async with get_client() as client:
-            work_pool = await client.read_work_pool(work_pool_name)
-    except ObjectNotFound:
-        raise RuntimeError(
-            f"Work pool {work_pool_name!r} does not exist. Please create it before"
-            " deploying this flow."
-        )
-
-    console = Console()
-    with Progress(
-        SpinnerColumn(),
-        TextColumn(f"Building image {full_image_name}..."),
-        transient=True,
-        console=console,
-    ) as progress:
-        docker_build_task = progress.add_task("docker_build", total=1)
-    auto_build = dockerfile == "auto"
-    build_kwargs["context"] = Path.cwd()
-    build_kwargs["tag"] = full_image_name
-    build_kwargs["pull"] = build_kwargs.get("pull", True)
-
-    if auto_build:
-        with generate_default_dockerfile():
-            build_image(**build_kwargs)
-    else:
-        build_kwargs["dockerfile"] = dockerfile
-        build_image(**build_kwargs)
-
-    progress.update(docker_build_task, completed=1)
-    console.print(f"Successfully built image {full_image_name!r}", style="green")
-
-    with Progress(
-        SpinnerColumn(),
-        TextColumn("Pushing image..."),
-        transient=True,
-        console=console,
-    ) as progress:
-        docker_push_task = progress.add_task("docker_push", total=1)
-
-        with docker_client() as client:
-            events = client.api.push(
-                repository=image_name, tag=tag, stream=True, decode=True
-            )
-            for event in events:
-                if "error" in event:
-                    raise PushError(event["error"])
-
-        progress.update(docker_push_task, completed=1)
-
-    console.print(f"Successfully pushed image {full_image_name!r}", style="green")
-
-    for deployment in track(
-        args,
-        description="Creating/updating deployments...",
-        console=console,
-    ):
-        await deployment.apply(image=full_image_name, work_pool_name=work_pool_name)
-
-    console.print("Successfully created/updated all deployments!\n", style="green")
-
-    table = Table(title="Deployments", show_header=False)
-
-    table.add_column(style="blue", no_wrap=True)
-
-    for deployment in args:
-        table.add_row(f"{deployment.flow_name}/{deployment.name}")
-
-    console.print(table)
-
-    if not work_pool.is_push_pool:
-        console.print(
-            "\nTo execute flow runs from these deployments, start a worker in a"
-            " separate terminal that pulls work from the"
-            f" {work_pool_name!r} work pool:"
-        )
-        console.print(
-            f"\n\t$ prefect worker start --pool {work_pool_name!r}",
-            style="blue",
-        )
-    console.print(
-        "\nTo trigger any of these deployments, use the"
-        " following command:\n[blue]\n\t$ prefect deployment run"
-        " [DEPLOYMENT_NAME]\n[/]"
-    )
-
-    if PREFECT_UI_URL:
-        console.print(
-            "\nYou can also trigger your deployments via the Prefect UI:"
-            f" [blue]{PREFECT_UI_URL.value()}/deployments[/]\n"
-        )
